Fix issues in the initial DEFAULTS file and version 7 migration script
First, embedding it in the Python script was stupid as \ characters
got mangled, it's now in a separate file.

The strip commands were totally broken, causing certain builds to fail,
which is fixed now.

The default commands for the 'cpan' build system are now in line with
those built into Morph.

The 'autotools' commands have been tweaked too, so that now they produce
identical cache keys to the defaults that are built into Morph (as of
morph.git commit 07df32fbd57477e5808cdbace965edcd0a81348f). This means
that moving to definitions format version 7 should not trigger any
rebuilds of anything.

Also, I added the 'module-build' build system (which was added to Morph
in commit f6613fe1ee6d879192fd4e503cb632b0dcab1fe7), so that it can be
used in the reference systems once they use definitions format version 7.

The migration script will warn if the definitions are version 7 already
but there is no DEFAULTS file, because that's unlikely to be deliberate.

Change-Id: I0a739ef38f521530e0d86d7330d1bcecf0a5bb73